refactor generation pipeline services
