+++ packages/server/src/services/generation/game-party-utils.ts
@@ -1,0 +1,19 @@
+function normalizePartyLookupName(value: string): string {
+  return value
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, " ")
+    .trim();
+}
+
+export function buildPartyNpcId(name: string): string {
+  const slug = normalizePartyLookupName(name).replace(/\s+/g, "-");
+  const encodedSlug = encodeURIComponent(name.trim().toLowerCase())
+    .replace(/%/g, "")
+    .replace(/[^a-z0-9]+/g, "-")
+    .replace(/(^-|-$)/g, "");
+  return `npc:${slug || encodedSlug || "unknown"}`;
+}
+
+export function isPartyNpcId(id: string): boolean {
+  return id.startsWith("npc:");
+}
