+++ packages/server/src/services/generation/spotify-tool-availability.ts
